refactor: response dto query 패키지로 이동
