+++ backend/baton/src/main/java/touch/baton/domain/member/query/controller/response/RunnerMyProfileResponse.java
@@ -1,6 +1,6 @@
 package touch.baton.domain.member.query.controller.response;
 
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponse;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
 
 import java.util.List;
