+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/query/controller/RunnerPostQueryController.java
@@ -21,10 +21,10 @@ import touch.baton.domain.oauth.query.controller.resolver.AuthMemberPrincipal;
 import touch.baton.domain.oauth.query.controller.resolver.AuthRunnerPrincipal;
 import touch.baton.domain.oauth.query.controller.resolver.AuthSupporterPrincipal;
 import touch.baton.domain.runnerpost.command.RunnerPost;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponse;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponses;
-import touch.baton.domain.runnerpost.command.controller.response.SupporterRunnerPostResponse;
-import touch.baton.domain.runnerpost.command.controller.response.SupporterRunnerPostResponses;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponses;
+import touch.baton.domain.runnerpost.query.controller.response.SupporterRunnerPostResponse;
+import touch.baton.domain.runnerpost.query.controller.response.SupporterRunnerPostResponses;
 import touch.baton.domain.runnerpost.command.vo.ReviewStatus;
 import touch.baton.domain.runnerpost.query.service.RunnerPostQueryService;
 import touch.baton.domain.runnerpost.query.service.dto.PageParams;
