+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/query/controller/response/RunnerPostResponses.java
@@ -1,4 +1,4 @@
-package touch.baton.domain.runnerpost.command.controller.response;
+package touch.baton.domain.runnerpost.query.controller.response;
 
 import java.util.List;
