+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/query/service/RunnerPostQueryService.java
@@ -11,8 +11,8 @@ import touch.baton.domain.member.query.controller.response.RunnerResponse;
 import touch.baton.domain.member.query.repository.SupporterRunnerPostQueryRepository;
 import touch.baton.domain.runnerpost.command.RunnerPost;
 import touch.baton.domain.runnerpost.command.RunnerPostsApplicantCount;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponse;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponses;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponses;
 import touch.baton.domain.runnerpost.command.exception.RunnerPostBusinessException;
 import touch.baton.domain.runnerpost.command.repository.dto.RunnerPostApplicantCountDto;
 import touch.baton.domain.runnerpost.command.vo.ReviewStatus;
