+++ backend/baton/src/test/java/touch/baton/assure/runnerpost/support/RunnerPostPageSupport.java
@@ -7,8 +7,8 @@ import touch.baton.assure.common.AssuredSupport;
 import touch.baton.assure.common.QueryParams;
 import touch.baton.domain.member.query.controller.response.RunnerResponse;
 import touch.baton.domain.runnerpost.command.RunnerPost;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponse;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponses;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponses;
 import touch.baton.domain.runnerpost.command.vo.ReviewStatus;
 
 import java.util.List;
