+++ backend/baton/src/test/java/touch/baton/assure/runnerpost/support/query/detail/RunnerPostDetailSupport.java
@@ -6,7 +6,7 @@ import touch.baton.assure.common.AssuredSupport;
 import touch.baton.assure.common.PathParams;
 import touch.baton.domain.member.command.Runner;
 import touch.baton.domain.member.query.controller.response.RunnerResponse;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponse;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
 import touch.baton.domain.runnerpost.command.vo.ReviewStatus;
 
 import java.time.LocalDateTime;
