+++ backend/baton/src/test/java/touch/baton/domain/runnerpost/command/service/RunnerPostQueryServiceTest.java
@@ -17,8 +17,8 @@ import touch.baton.domain.member.command.vo.MemberName;
 import touch.baton.domain.member.command.vo.OauthId;
 import touch.baton.domain.member.command.vo.SocialId;
 import touch.baton.domain.runnerpost.command.RunnerPost;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponse;
-import touch.baton.domain.runnerpost.command.controller.response.RunnerPostResponses;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
+import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponses;
 import touch.baton.domain.runnerpost.command.exception.RunnerPostBusinessException;
 import touch.baton.domain.runnerpost.command.vo.CuriousContents;
 import touch.baton.domain.runnerpost.command.vo.Deadline;
